Added to jmx api count of downloaded docs and count of discovered. Moved frontier and thread reports to be their own operations (Content is too large to be attributes).  Also, this commit should take care of jmx shutdown for sure. *  src/java/org/archive/crawler/Heritrix.java     Added logging of registration/deregistration of MBeans.     (registeredHeritrixObjectInstance): Keep around the registered heritrix     instance and then unregister it on shutdown (Allows heritrix to be     redeployed inside a container w/o having to restart container).     (getCurrentJob): Removed. Unused.     (stop): Call shutdown. * src/java/org/archive/crawler/admin/CrawlJob.java     Added DOWNLOAD_COUNT_ATTR and DISCOVERED_COUNT_ATTR.     Moved FrontiersReport and ThreadsReport to be operations instead of     attributes. * src/java/org/archive/crawler/admin/StatisticsTracker.java * src/java/org/archive/crawler/framework/StatisticsTracking.java     Moved javadoc to interface (Added two methods to ST interface,     totalCount and successfullyFetchedCount.

+++ heritrix/src/java/org/archive/crawler/admin/StatisticsTracker.java
@@ -362,7 +362,7 @@ implements CrawlURIDispositionListener {
         synchronized (map) {
             LongWrapper lw = (LongWrapper)map.get(key);
             if(lw == null) {
-                map.put(key, new LongWrapper(1));
+                map.put(key, new LongWrapper((long)1));
             } else {
                 lw.longValue += increment;
             }
@@ -558,16 +558,6 @@ implements CrawlURIDispositionListener {
         return shouldrun ? controller.getFrontier().disregardedUriCount() : downloadDisregards;
     }
 
-    /**
-     * Number of <i>successfully</i> processed URIs.
-     *
-     * <p>If crawl not running (paused or stopped) this will return the value
-     * of the last snapshot.
-     *
-     * @return The number of successully fetched URIs
-     *
-     * @see org.archive.crawler.framework.Frontier#succeededFetchCount()
-     */
     public long successfullyFetchedCount() {
         // While shouldrun is true we can use info direct from the crawler.
         // After that our last snapshot will have to do.
@@ -575,11 +565,6 @@ implements CrawlURIDispositionListener {
             controller.getFrontier().succeededFetchCount() : downloadedUriCount;
     }
     
-    /**
-     * Total number of URIs (processed + queued + currently being processed)
-     *  
-     * @return The total number of  
-     */
     public long totalCount() {
         return queuedUriCount() + activeThreadCount() +
             successfullyFetchedCount();
